ref: Move the init function to sentry crate
This is in preparation of moving the default options to the sentry
crate as well, which is needed to move the default transport and
integrations out of `sentry-core`.

Along with this, it removes some deprecated functions.
