+++ sentry/src/lib.rs
@@ -111,5 +111,26 @@
 //! * `with_test_support`: Enables the test support module.
 #![warn(missing_docs)]
 
+#[macro_use]
+extern crate sentry_core;
+
+#[cfg(feature = "with_client_implementation")]
+mod init;
+
 #[doc(inline)]
 pub use sentry_core::*;
+
+/// Useful internals.
+///
+/// This module contains types that users of the crate typically do not
+/// have to interface with directly.  These are often returned
+/// from methods on other types.
+pub mod internals {
+    pub use sentry_core::internals::*;
+
+    #[cfg(feature = "with_client_implementation")]
+    pub use crate::init::ClientInitGuard;
+}
+
+#[cfg(feature = "with_client_implementation")]
+pub use crate::init::init;
